Implement simple loop and change exit status rule
Implemented "loop block" extended from "do loop". The loop is
breaked when a command in the block gives a more than zero
exit status. The program goes to the next line that is under
the loop block.

Since I want to stop all procedure when a parse error occurs,
I have changed the rule of exit status on GlueLang. The detail
will be described in the other document.